requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
pymupdf==1.24.10
openai-agents>=0.0.5
openai>=1.50.0
python-dotenv==1.0.1
